Move fswatcher package out of crypto/certloader
Given that this logic is totally agnostic to the crypto/certloader
package and in a subsequent commit it will be used also in the IPSec
subsystem, make it a top level package.
