+++ pkg/crypto/certloader/watcher.go
@@ -20,7 +20,7 @@ import (
 
 	"github.com/sirupsen/logrus"
 
-	"github.com/cilium/cilium/pkg/crypto/certloader/fswatcher"
+	"github.com/cilium/cilium/pkg/fswatcher"
 	"github.com/cilium/cilium/pkg/logging/logfields"
 )
