+++ pkg/fswatcher/fswatcher.go
