+++ pkg/fswatcher/fswatcher_test.go
